+++ v2/index.d.ts
@@ -1,5 +1,0 @@
-// v2/index.d.ts
-
-export default function toPinyinTones(longString: string): string;
-export function unspacedPinyinSplit(text: string): string[];
-export function convertUnspacedPinyin(unspacedLongString: string): string;